refactor(graduate): 중복 코드 수정
