--- apps/graduate/src/pages/admin/all/ui/UserDetailModal/index.ts
@@ -1,5 +1,0 @@
-export { default as UserApplicationView } from './UserApplicationView';
-export { default as UserDetailView } from './UserDetailView';
-export { default as UserMiddleReportView } from './UserMiddleReportView';
-export { default as UserFinalReportView } from './UserFinalReportView';
-export { default as UserDetailModal } from './UserDetailModal';
